Start moving code into clode/ directory
To support Python wheels, the code needs
to be in a subdirectory named after the Python package.
Otherwise, its files get spread all over the python
libraries folder.

This modifies a great number of paths, in particular
the .cl include paths, so the path inclusions
need to be fixed and tested.